+++ src/main/java/com/github/pdfhandler4j/imp/ByPagesPdfSplitter.java
@@ -12,7 +12,7 @@ public class ByPagesPdfSplitter extends ByVolumePdfSplitter{
   }
   
   @Override
-  protected boolean mustSplit(long currentCombined, PageRange range, long max, int totalPages) {
+  protected boolean mustSplit(long currentCombined, PageRange range, long max, long totalPages) {
     return currentCombined >= range.end();
   };
